Preserve metadata across platforms
